itest: move balance related tests into one file
